[client] | refactor(TagVideos): 태그 비디오 페이지 요구사항 반영
훅을 분리하였고, 명확하지 않은 코드를 변경하였습니다.
