+++ packages/client/views/TagVideos/index.tsx
@@ -3,48 +3,22 @@ import Grid from '@material-ui/core/Grid';
 import { useRouter } from 'next/router';
 
 import * as S from './style';
-import { sortOptions, endpoint } from '../../constants';
 import TagsSVG from '../../svgs/TagsSVG';
+import { sortOptions } from '../../constants';
 import Layout from '../../components/Layout';
 import VideoItem from '../../components/VideoItem';
-import CircularProgress from '../../components/CircularProgress';
-import { NATURAL_NUMBER_REGEX } from '../../libs/regex';
-import { useTag } from './hook/use-tag';
 import { useTagVideos } from './hook/use-tag-videos';
+import CircularProgress from '../../components/CircularProgress';
 
 const TagVideos: React.FunctionComponent = () => {
-  const router = useRouter();
-  const { tagId } = router.query;
-  const validatedTagId = NATURAL_NUMBER_REGEX.test(tagId.toString())
-    ? Number(tagId)
-    : null;
-
-  const [activeSortOption, setActiveSortOption] = useState(
-    sortOptions[0].value,
-  );
-  const [page, setPage] = useState(1);
-
-  const { tag, error } = useTag(validatedTagId);
-  const { videos, hasMore } = useTagVideos(
-    validatedTagId,
-    page,
+  const {
+    tag,
+    videos,
+    handleFilterClick,
+    handlePageChange,
     activeSortOption,
-  );
-
-  useEffect(() => {
-    if (!tagId || error) {
-      router.push(endpoint.hotlist);
-    }
-  }, [error]);
-
-  const handleFilterClick = value => {
-    setActiveSortOption(value);
-    setPage(1);
-  };
-
-  const handlePageChange = () => {
-    setPage(page + 1);
-  };
+    hasMore,
+  } = useTagVideos();
 
   return (
     <Layout>
